Company Database System (miniproject)
The Company Database System is a distributed application, developed to maintain the details of employees who applied for new job. It maintains the information about the personal details of the employees. This application is helpful to department of the organization which maintains data of employees related to an organization. We can also view the details of employee who applied for the particular job. We can also add and remove the details of employee in the table.

Operating System                                          :   Windows 10,7
Back End Language                                         :   PHP
Front End Language                                        :   HTML, CSS
Database                                                  :   MySQL
Database Management tool                                  :   Xampp
Integrated Development Environment                        :   Sublime Text Editor
To run this project we need to install xampp server and need to create company database (I have given name as company).In that I have created 4Tables namely app_development,data_scientist,signin,web_developer(I recommend to create these tables before running this code in your system and be cautious when you are giving the name of the tables in your code and in database as it is case-sensitive).final project's output photos is given the folder "project_output_pics".